refactor: Separate code into modules
