+++ src/todos/addTodoToList.js
@@ -1,0 +1,6 @@
+const addTodoToList = (todo, list) => {
+  todo.list = list.title
+  list.todos.push(todo)
+}
+
+export default addTodoToList
