+++ src/todos/changeTodoPriority.js
@@ -1,0 +1,5 @@
+const changeTodoPriority = (todo, newPriority) => {
+  todo.priority = newPriority
+}
+
+export default changeTodoPriority
