+++ src/todos/completeTodo.js
@@ -1,0 +1,3 @@
+const completeTodo = (todo) => (todo.complete = true)
+
+export default completeTodo
